SpaceAppFilesModelFactory
Still un-integrated, but this is hopefully enough to get design
discussion rolling.

This adds building the SpaceApp model object from a directory hierarchy:
- systemd files are loaded (no more annoying\nJSON\nencoding!)
- environment files are loaded (though these weren't as bad in JSON)
- other files are bundled (i.e. nginx.conf, cert.pem, etc)

Each of the above can be overwritten in a particular region using
subdirectories.
`/us-east-1/app.json` is automatically nested as the `us-east-1` child of
`/app.json`.

I'm keeping the bias that deployment metadata is best
kept/controlled/reviewed with application code: using a
`.space` hidden directory to isolate this data.